Add directory color support for GNU ls
GNU ls is part of the coreutils package. It's output, if color is enabled, can
be tuned to the badwolf colorscheme with the following command:

    $ eval $(dircolors -b contrib/badwolf.dircolors)

To make this the default, copy it to ~/.dircolors.  Many systems are already
configured to look for this file, but you may need to add something like:

    $ eval $(dircolors -b ~/.dircolors)

to ~/.bashrc.  To enable color support in ls, you could do the following:

    $ alias ls='ls --color=auto'